libm: Fixed the sign mask macro in vrd2 & 4 tan()
vrd2_tan() & vrd4_tan() reported accuracy issues on Windows platform.
This was because the sign mask macro, defined as (1UL << 63),
caused issues on Windows where UL was considered as 32-bit.

To resolve this issue, the sign mask macro is replaced with a
pre-computed uint64_t value, specifically 0x8000000000000000.

Change-Id: Ifb75744971590c3daf05eaf0a26063ade23c6ee3
Signed-off-by: Kiran M Vijay <kiran.mvijay@amd.com>
AMD-Internal: [CPUPL-5369]