libm: Fixed the sign mask macro tan APIs
vrd8_tan() reported accuracy issues on Windows platform.
This was because the sign mask macro, defined as (1UL << 63),
caused issues on Windows where UL was considered as 32-bit.

To resolve this issue, the sign mask macro is replaced with a
pre-computed uint64_t value, specifically 0x8000000000000000.

This change is applied in fast/tan variant as well.

Change-Id: I85af5546dbf8956b3b92814055e8760318d22813
Signed-off-by: Kiran M Vijay <kiran.mvijay@amd.com>
AMD-Internal: [CPUPL-5603]