libm: fmod algorithm changed to improve accuracy
The fmod double variant previously used the same logic as the
single variant, leading to high ULP error in certain corner cases.

To resolve this, the assembly variant logic of fmod has been optimized and converted to C.
This optimization improves the accuracy and precision of the fmod function.

MAX ULP for the reported corner cases = 0

Change-Id: I370722ead4742110652838b75fcf97c411eecf91
Signed-off-by: Kiran M Vijay <kiran.mvijay@amd.com>
AMD-Internal: [CPUPL-5687]