CMake:Updated aocl-libm cmake for static-dispatch
For static dispatch, when ALM_STATIC_DISPATCH options
matches a partial keyword, configuration was succesful.
for example:
$cmake --preset dev-release-gcc -DALM_STATIC_DISPATCH=AVX --fresh
static dispatch was set for AVX2

so updated aocl-libm cmakelists.txt for static dispatch with
keyword "STREQUAL" instead of "MATCHES" and only configures
when valid options are given, otherwise throw an error message.

Signed-off-by: Jagadish R <jagadish1.r@amd.com>
Change-Id: I7b5e5b4f7e7e212aa2530ed26c1b9d6deec7ef54