libm:src:nextafterf: Bug fix for nextafterf(from,to)
Added Condition:
An additional condition to check if the output is infinity. Previously, NaN was being returned even when the output was infinity.

Special Case Handling for from=0 and to!=from:
Fixed to return the correct value (+/- 0) and raise the exceptions FE_INEXACT and FE_UNDERFLOW.

Special Cases Handling:
Replaced __alm_handle_error calls in the source code with alm_nextafterf_special

Signed-off-by: Mangala V <mangala.v@amd.com
Change-Id: I39b7dbaddb1faec38159b2b8e1449c7140a87d5c
AMD-Internal: [CPUPL-6313] [SWLCSG-3019]