libm:cmake: Utils : Enhanced CMake to link AOCL-Utils static library
Modified libm utils cmake :
Added checks to determine if the utils library is present.
If found, link it with libm.
Added checks to determine if the utils directory is present.
If found, configure, build, and link it with libm.
Implemented cloning of aocl-utils from the specified branch,
followed by configuration, building, and linking with libm.

The aocl-utils will be placed within the build directory under
the external folder.

Signed-off-by: Jagadish R jagadish1.r@amd.com
AMD-Internal: [CPUPL-4243]
Change-Id: If2d8092073496bc8e9935bee7d3afae1cd44eb17