libm:src:exp10: Fixing indexing issue in single-precision vector array variant for in-place operation
When size of array is not a multiple of 4, results of some values
were being re-computed for the sake of performance. But this causes errors when the operation is
in-place (input and output arrays are the same)

Updated the code to save last 4 input values before looping through
remaining elements.

Signed-off-by: Nimmy Krishnan <nimmy.krishnan@amd.com>
AMD-Internal:[CPUPL-7631]