src:iface: Update ZEN, ZEN2, ZEN3 function table
Updated function table for ZEN, ZEN2, and ZEN3 architectures
with vrd8 and vrs16 variant functions. By design, this was kept
NULL function-pointer because ZEN, ZEN2, ZEN3 do not support
AVX512 ISA. But in some cases due to wrong prediction of hardware,
vrd8 and vrs16 variants are dispatched wrongly and making a function
call using NULL pointer. This results into crash.

To mitigate this issue, we are initializing function table with
respective functions so it will fail with "Illegal Instruction".

AMD-Internal:[CPUPL-7718]
Signed-off-by: Dhiraj Kumar <dhiraj.kumar@amd.com>